# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job-Search Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get list of companies",
                "description": "Returns list of companies",
                "operationId": "Get All Companies",
                "responses": {
                    "200": {
                        "description": "All companies"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/companies/show": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "show specific company",
                "description": "show specific company",
                "operationId": "show specific company",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "query",
                        "description": "Enter company id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company details"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/companies/create": {
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Create New Company",
                "description": "Add New Company",
                "operationId": "create new company",
                "requestBody": {
                    "description": "Pass companies credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "comapny"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "comapny@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Company created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/companies/edit": {
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Update company",
                "description": "Edit company",
                "operationId": "update company",
                "requestBody": {
                    "description": "Pass company credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "company_id": {
                                        "type": "integer",
                                        "format": "company_id",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "comapny"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "comapny@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Company update successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/companies/delete": {
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Soft delete company",
                "description": "Soft delete company",
                "operationId": "delete specific company",
                "requestBody": {
                    "description": "Pass role credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "company_id"
                                ],
                                "properties": {
                                    "company_id": {
                                        "type": "integer",
                                        "format": "company_id",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Company deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/companies/restore": {
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Restore delete comapny",
                "description": "restore comapny",
                "operationId": "restore specific company",
                "requestBody": {
                    "description": "Pass comapny credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "company_id"
                                ],
                                "properties": {
                                    "company_id": {
                                        "type": "integer",
                                        "format": "company_id",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Company restored successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}